+++ composeApp/src/androidMain/kotlin/com/cairosquad/evolvefit/repository/remote/provideHttpClient.android.kt
@@ -1,13 +1,15 @@
-package com.cairosquad.evolvefit.remote
+package com.cairosquad.evolvefit.repository.remote
 
-import com.cairosquad.evolvefit.local.AuthPreferences
+
+import com.cairosquad.evolvefit.repository.local.AuthPreferences
 import io.ktor.client.HttpClient
-import io.ktor.client.engine.darwin.Darwin
 import io.ktor.client.plugins.auth.Auth
 import io.ktor.client.plugins.auth.providers.BearerTokens
 import io.ktor.client.plugins.auth.providers.bearer
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.client.plugins.defaultRequest
+import io.ktor.client.plugins.logging.LogLevel
+import io.ktor.client.plugins.logging.Logging
 import io.ktor.serialization.kotlinx.json.json
 import kotlinx.serialization.json.Json
 
@@ -15,13 +17,16 @@ actual fun provideHttpClient(
     authPreferences: AuthPreferences,
     refreshTokenProvider: suspend (String) -> BearerTokens?
 ): HttpClient {
-    return HttpClient(Darwin) {
+    return HttpClient {
         install(ContentNegotiation) {
             json(Json { ignoreUnknownKeys = true })
         }
+        install(Logging) {
+            level = LogLevel.ALL
+        }
         defaultRequest {
             url("https://evolve-fit-dev.the-chance.net/")
-                  }
+        }
         install(Auth) {
             bearer {
                 loadTokens {
@@ -39,4 +44,3 @@ actual fun provideHttpClient(
         }
     }
 }
-
