+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/di/dataModule.kt
@@ -1,6 +1,6 @@
 package com.cairosquad.evolvefit.di
 
-import com.cairosquad.evolvefit.local.AuthPreferences
+import com.cairosquad.evolvefit.repository.local.AuthPreferences
 import com.russhwolf.settings.Settings
 import org.koin.dsl.module
