+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/di/remoteModule.kt
@@ -1,7 +1,7 @@
 package com.cairosquad.evolvefit.di
 
-import com.cairosquad.evolvefit.remote.workout.WorkoutRemoteDataSourceImpl
 import com.cairosquad.evolvefit.repository.remote.workout.WorkoutRemoteDataSource
+import com.cairosquad.evolvefit.repository.remote.workout.WorkoutRemoteDataSourceImpl
 import org.koin.dsl.module
 
 val remoteModule = module {
