+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/domain/entity/Equipment.kt
@@ -1,0 +1,6 @@
+package com.cairosquad.evolvefit.domain.entity
+
+data class Equipment(
+    val id: Long,
+    val name: String
+)
