+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/domain/entity/Gender.kt
@@ -1,0 +1,6 @@
+package com.cairosquad.evolvefit.domain.entity
+
+enum class Gender {
+    MALE,
+    FEMALE
+}
