+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/domain/entity/Workout.kt
@@ -1,4 +1,4 @@
-package com.cairosquad.evolvefit.entity
+package com.cairosquad.evolvefit.domain.entity
 
 
 data class Workout(
