+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/domain/entity/WorkoutDay.kt
@@ -1,4 +1,4 @@
-package com.cairosquad.evolvefit.entity
+package com.cairosquad.evolvefit.domain.entity
 
 enum class WorkoutDay {
     MONDAY,
