+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/domain/usecase/workout/ExerciseUseCase.kt
@@ -1,13 +1,13 @@
 package com.cairosquad.evolvefit.domain.usecase.workout
 
+import com.cairosquad.evolvefit.domain.entity.Equipment
+import com.cairosquad.evolvefit.domain.entity.Exercise
 import com.cairosquad.evolvefit.domain.repository.WorkoutRepository
-import com.cairosquad.evolvefit.entity.Exercise
-import com.cairosquad.evolvefit.entity.Tool
 
 class ExerciseUseCase(
     private val workOutRepository: WorkoutRepository
 ) {
-    suspend fun getEquipments(): List<Tool> {
+    suspend fun getEquipments(): List<Equipment> {
         return workOutRepository.getEquipments()
     }
