+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/domain/usecase/workout/ManageWorkoutUseCase.kt
@@ -1,7 +1,7 @@
 package com.cairosquad.evolvefit.domain.usecase.workout
 
+import com.cairosquad.evolvefit.domain.entity.Workout
 import com.cairosquad.evolvefit.domain.repository.WorkoutRepository
-import com.cairosquad.evolvefit.entity.Workout
 
 class ManageWorkoutUseCase(
     private val workoutRepository: WorkoutRepository
