--- composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/entity/Tool.kt
@@ -1,6 +1,0 @@
-package com.cairosquad.evolvefit.entity
-
-data class Tool(
-    val id: Long,
-    val name: String
-)
