+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/repository/remote/auth/AuthRemoteDataSourceImp.kt
@@ -1,12 +1,7 @@
-package com.cairosquad.evolvefit.remote.auth
+package com.cairosquad.evolvefit.repository.remote.auth
 
-import com.cairosquad.evolvefit.remote.safeApiCall
 import com.cairosquad.evolvefit.repository.remote.AuthRemoteDataSource
-import com.cairosquad.evolvefit.repository.remote.auth.AuthResponse
-import com.cairosquad.evolvefit.repository.remote.auth.EquipmentDto
-import com.cairosquad.evolvefit.repository.remote.auth.LoginRequest
-import com.cairosquad.evolvefit.repository.remote.auth.RefreshRequest
-import com.cairosquad.evolvefit.repository.remote.auth.RegisterRequest
+import com.cairosquad.evolvefit.repository.remote.safeApiCall
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
