+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/repository/remote/workout/WorkoutRemoteDataSourceImpl.kt
@@ -1,8 +1,6 @@
-package com.cairosquad.evolvefit.remote.workout
+package com.cairosquad.evolvefit.repository.remote.workout
 
 import com.cairosquad.evolvefit.repository.remote.auth.EquipmentDto
-import com.cairosquad.evolvefit.repository.remote.workout.ExerciseDto
-import com.cairosquad.evolvefit.repository.remote.workout.WorkoutRemoteDataSource
 
 class WorkoutRemoteDataSourceImpl: WorkoutRemoteDataSource {
     override suspend fun getEquipments(): List<EquipmentDto> {
